Gut Contact module, move into its own directory, return vectors instead of packing.
